\name{msBP.internal}
\alias{msBP.clus2prob}
\alias{msBP.leaf.allocation}
\alias{msBP.marginalBeta}
\alias{msBP.nesting}
\alias{msBP.nrvTree}
\alias{msBP.posteriorH0}
\title{
Internal msBP functions
}
\description{
Internal msBP functions
}
\details{
These are not to be called by the user.
}
\keyword{ msBP }
\keyword{ internal}
