\name{msBP-package}
\alias{msBP-package}
\alias{msBP}
\docType{package}
\title{
Bayesian nonparametric density estimation via Multiscale Bernstein Polynomials (msBP)
}
\description{
Performs Bayesian nonparametric multiscale density estimation and multiscale testing of group differences with multiscale Bernstein polynomials (msBP) mixtures as in Canale and Dunson (2016).
}
\details{
\tabular{ll}{
Package: \tab msBP\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2016-04-19\cr
License: \tab GNU GPL\cr
}
}
\author{
Antonio Canale <antonio.canale@unito.it>
}
\references{
Canale, A. and Dunson, D. B. (2016), "Multiscale Bernstein polynomials for densities", Statistica Sinica, 26
}
\keyword{Binary trees, Density estimation, Multiscale stick-breaking, Multiscale testing}
