\name{msBP-package}
\alias{msBP-package}
\alias{msBP}
\docType{package}
\title{
Bayesian nonparametric density estimation via Multiscale Bernstein Polynomials (msBP)
}
\description{
Performs Bayesian nonparametric multiscale density estimation and multiscale testing of group differences with multiscale Bernstein polynomials (msBP) mixtures as in Canale and Dunson (2016).
}
\author{
Antonio Canale <antonio.canale@unipd.it>
}
\references{
Canale, A. and Dunson, D. B. (2016), "Multiscale Bernstein polynomials for densities", Statistica Sinica, 26(3), 1175-1195.

Canale, A. (2017), "msBP: An R Package to Perform Bayesian Nonparametric Inference Using Multiscale Bernstein Polynomials Mixtures". Journal of Statistical Software, 78(6), 1-19. 
}
\keyword{Multiscale stick-breaking}
