\name{msBP-package}
\alias{msBP-package}
\alias{msBP}
\docType{package}
\title{
Bayesian nonparametric density estimation via Multiscale Bernstein Polynomials (msBP)
}
\description{
Performs Bayesian nonparametric multiscale density estimation and multiscale testing of group differences with multiscale Bernstein polynomials (msBP) mixtures as in Canale and Dunson (2016).
}
\details{
\tabular{ll}{
Package: \tab msBP\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2017-06-06\cr
License: \tab GNU GPL\cr
}
}
\author{
Antonio Canale <canale@stat.unipd.it>
}
\references{
Canale, A. and Dunson, D. B. (2016), "Multiscale Bernstein polynomials for densities", Statistica Sinica, 26(3), 1175-1195.

Canale, A. (2017), "msBP: An R Package to Perform Bayesian Nonparametric Inference Using Multiscale Bernstein Polynomials Mixtures". Journal of Statistical Software, 78(6), 1-19. 
}
\keyword{Binary trees, Density estimation, Multiscale stick-breaking, Multiscale testing}
